+++ packages/react-router/__tests__/server-runtime/handle-error-test.ts
@@ -1,33 +1,28 @@
-import type { ServerBuild } from "../../lib/server-runtime/build";
 import { createRequestHandler } from "../../lib/server-runtime/server";
 import { ErrorResponseImpl } from "../../lib/router/utils";
+import { mockServerBuild } from "./utils";
+import type { HandleDocumentRequestFunction } from "../../lib/server-runtime/build";
 
-function getHandler(routeModule = {}, entryServerModule = {}) {
-  let routeId = "root";
+function getHandler(
+  routeModule = {},
+  opts: {
+    handleDocumentRequest?: HandleDocumentRequestFunction;
+  } = {}
+) {
   let handleErrorSpy = jest.fn();
-  let build = {
-    routes: {
-      [routeId]: {
-        id: routeId,
+  let build = mockServerBuild(
+    {
+      root: {
         path: "/",
-        module: {
-          default() {},
-          ...routeModule,
-        },
+        default() {},
+        ...routeModule,
       },
     },
-    entry: {
-      module: {
-        handleError: handleErrorSpy,
-        default() {
-          return new Response("<html><body>Dummy document</body></html>");
-        },
-        ...entryServerModule,
-      },
-    },
-    future: {},
-    prerender: [],
-  } as unknown as ServerBuild;
+    {
+      handleError: handleErrorSpy,
+      handleDocumentRequest: opts.handleDocumentRequest,
+    }
+  );
 
   return {
     handler: createRequestHandler(build),
@@ -76,7 +71,7 @@ describe("handleError", () => {
 
     it("provides render-thrown Error", async () => {
       let { handler, handleErrorSpy } = getHandler(undefined, {
-        default() {
+        handleDocumentRequest() {
           throw new Error("Render error");
         },
       });
@@ -112,7 +107,7 @@ describe("handleError", () => {
           throw error;
         },
       });
-      let request = new Request("http://example.com/?_data=root");
+      let request = new Request("http://example.com/_root.data");
       await handler(request);
       expect(handleErrorSpy).toHaveBeenCalledWith(error, {
         request,
@@ -123,7 +118,7 @@ describe("handleError", () => {
 
     it("provides router-thrown ErrorResponse", async () => {
       let { handler, handleErrorSpy } = getHandler({});
-      let request = new Request("http://example.com/?_data=root", {
+      let request = new Request("http://example.com/_root.data", {
         method: "post",
       });
       await handler(request);
@@ -153,9 +148,64 @@ describe("handleError", () => {
           );
         },
       });
-      let request = new Request("http://example.com/?_data=root");
+      let request = new Request("http://example.com/_root.data");
       await handler(request);
       expect(handleErrorSpy).not.toHaveBeenCalled();
+    });
+
+    it("provides proper params to handleError", async () => {
+      let error = new Error("💥");
+
+      let handleErrorSpy = jest.fn();
+      let build: ServerBuild = {
+        routes: {
+          param: {
+            id: "param",
+            path: "/:param",
+            module: {
+              default() {
+                return null;
+              },
+              loader() {
+                throw error;
+              },
+            },
+          },
+        },
+        entry: {
+          module: {
+            handleError: handleErrorSpy,
+            default() {
+              return new Response("<html><body>Dummy document</body></html>");
+            },
+          },
+        },
+        future: {
+          // Fill in the required values
+          unstable_middleware: false,
+          unstable_subResourceIntegrity: false,
+        },
+        prerender: [],
+        assets: {
+          entry: { imports: [], module: "" },
+          routes: {},
+          url: "",
+          version: "",
+        },
+        assetsBuildDirectory: "",
+        publicPath: "/",
+        ssr: true,
+        isSpaMode: false,
+      };
+
+      let handler = createRequestHandler(build);
+      let request = new Request("http://example.com/a.data");
+      await handler(request);
+      expect(handleErrorSpy).toHaveBeenCalledWith(error, {
+        request,
+        params: { param: "a" },
+        context: {},
+      });
     });
   });