+++ packages/react-router/lib/dom/ssr/entry.ts
@@ -3,6 +3,7 @@ import type { StaticHandlerContext } from "../../router/router";
 import type { EntryRoute } from "./routes";
 import type { RouteModules } from "./routeModules";
 import type { RouteManifest } from "../../router/utils";
+import type { ServerBuild } from "../../server-runtime/build";
 
 type SerializedError = {
   message: string;
@@ -18,6 +19,7 @@ export interface FrameworkContextObject {
   future: FutureConfig;
   ssr: boolean;
   isSpaMode: boolean;
+  routeDiscovery: ServerBuild["routeDiscovery"];
   serializeError?(error: Error): SerializedError;
   renderMeta?: {
     didRenderScripts?: boolean;